xmllite/writer: Maintain a stack of written elements to write end tags.
Xmllite helps in a way that user doesn't have to specify closing tag
name (like it has to in case of MXWriter). So when closing current level
element qname is used for a closing tag; later this stack will also hold
namespace definitions that element is carrying to check if nested
element uses defined prefix.